Agent

+++ AgriHub/backend/src/main/java/com/backend/AgriLearn/repositories/UserRepository.java
@@ -5,8 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-//    User findByEmail(String email);
-//    User findByUsername(String username);
-//    User findByUsernameOrEmail(String username, String email);
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
